requirements.txt
requests
pandas
py-clob-client
python-dotenv
web3
